+++ News.md
@@ -3,9 +3,9 @@
 ## New APIs
 
 - `Language#name`.
+- `Language#name` returns the name of the parser, a string.
 - `Language#metadata` returns a `LanguageMetadata` object with
   `major_version`, `minor_version`, and `patch_version` fields.
-- `Language#name` returns the name of the parser, a string.
 
 ## Breaking Changes